Small tightening up of type casting to avoid a compiler warning

+++ MusicKit/Frameworks/SndKit/SndFunctions.m
@@ -1191,8 +1191,8 @@ int SndRead(FILE *fp, SndSoundStruct **sound, const char *fileTypeStr)
 
 	if([[NSUserDefaults standardUserDefaults] boolForKey: @"SndShowInputFileFormat"]) {
             printf("Input file: using sample rate %lu Hz, size %s, style %s, %d %s\n",
-               informat.info.rate, st_sizes_str[informat.info.size],
-               st_encodings_str[informat.info.encoding], informat.info.channels,
+               informat.info.rate, st_sizes_str[(int)(informat.info.size)],
+               st_encodings_str[(int)(informat.info.encoding)], informat.info.channels,
                (informat.info.channels > 1) ? "channels" : "channel");
             if (informat.comment)
                 printf("Input file: comment \"%s\"\n", informat.comment);
